feat. rank 자료형 변경

+++ src/main/java/com/simulation/reactgame/RecordRankingView.java
@@ -4,5 +4,5 @@ public interface RecordRankingView {
     Long getId();
     String getName();
     float getAverageTime();
-    Integer getRank();
+    Long getRank();
 }
